+++ src/agenda/graphic.py
@@ -1,7 +1,7 @@
 from src._road.road import get_parent_road, RoadUnit, is_sub_road
 from src.agenda.idea import IdeaUnit
 from src.agenda.agenda import AgendaUnit
-from src.agenda.report import (
+from src.agenda.report_agenda import (
     get_agenda_partyunits_dataframe,
     get_agenda_intent_dataframe,
 )
