+++ src/world/test_world/test_world_person_kpi.py
@@ -1,4 +1,4 @@
-from src.world.report import (
+from src.world.report_world import (
     get_world_guts_partys_dataframe,
     get_world_guts_partys_plotly_fig,
     get_world_lives_partys_dataframe,
